GEODE-4408: Move CacheableKeys.hpp to internal include dir
